requirements.txt
django==3.2.15
django-debug-toolbar==3.2.1
Pillow==8.2.0
environs[django]==9.3.2
djangorestframework==3.14.0
psycopg2==2.9.5
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
requests==2.28.2
geopy==2.3.0
